backend/requirements.txt
Django>=5.1.2,<5.2
djangorestframework>=3.15.2,<3.16
